Restructure Pfs module
